requirements.txt
tensorflow
numpy
pillow
ibm-watson-machine-learning
flask